requirements.txt
pandas
requests
pyspark
sqlalchemy
psycopg2-binary


- Install the dependencies by running:

  ```bash
  pip install -r requirements.txt



